requirements.txt
fastapi[standard]
SQLAlchemy
python-dotenv
pymysql
cryptography
bcrypt
jose